+++ packages/abap-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/abap-deploy-config-inquirer
 
+## 2.0.26
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/fiori-generator-shared 1.1.0 → 1.2.0
+- @sap-ux/inquirer-common 1.0.22 → 1.0.23
+
+## 2.0.25
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/fiori-generator-shared 1.0.20 → 1.1.0
+- @sap-ux/inquirer-common 1.0.21 → 1.0.22
+
 ## 2.0.24
 
 ### Patch Changes