+++ packages/backend-proxy-middleware-cf/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux/backend-proxy-middleware-cf
 
+## 0.3.33
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.32 → 1.0.33
+
+## 0.3.32
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.31 → 1.0.32
+
+## 0.3.31
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.30 → 1.0.31
+
 ## 0.3.30
 
 ### Patch Changes