+++ packages/deploy-tooling/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/deploy-tooling
 
+## 1.0.24
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/inquirer-common 1.0.22 → 1.0.23
+
+## 1.0.23
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/inquirer-common 1.0.21 → 1.0.22
+
 ## 1.0.22
 
 ### Patch Changes